requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask == 1.1.2
Pillow == 8.4.0
imageio == 2.9.0
keras == 2.9.0
matplotlib == 3.4.3
numpy == 1.20.3
opencv_python == 4.6.0.66
plotly == 5.10.0
scikit_learn == 0.24.2
scipy == 1.7.1
brisque == 0.0.13
tensorflow == 2.9.1
